// src/shared/dependencies/address.pb.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.5
//   protoc               v6.32.0
// source: address.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "address";

export interface Address {
  addressId?: string | undefined;
  street?: string | undefined;
  building?: string | undefined;
  apartment?: string | undefined;
  district?: string | undefined;
  city?: string | undefined;
  state?: string | undefined;
  postalCode?: string | undefined;
  country?: string | undefined;
  landmark?: string | undefined;
  directionUrl?: string | undefined;
  latitude?: number | undefined;
  longitude?: number | undefined;
}

export interface CreateAddressRequest {
  entityType: string;
  entityId: string;
  street: string;
  building: string;
  apartment: string;
  district: string;
  city: string;
  state: string;
  postalCode: string;
  country: string;
  landmark: string;
  directionUrl: string;
  latitude: number;
  longitude: number;
}

export interface VerifyAddressRequest {
  entityType: string;
  entityId: string;
}

export interface CreateAddressResponse {
  address: Address | undefined;
  status: boolean;
}

export interface AddressResponse {
  address: Address | undefined;
  status: boolean;
}

export const ADDRESS_PACKAGE_NAME = "address";

/** 📁 Files Service - handles file uploads, storage, and management */

export interface AddressServiceClient {
  verifyAddress(request: VerifyAddressRequest): Observable<AddressResponse>;

  createOrUpdateAddress(request: CreateAddressRequest): Observable<AddressResponse>;
}

/** 📁 Files Service - handles file uploads, storage, and management */

export interface AddressServiceController {
  verifyAddress(
    request: VerifyAddressRequest,
  ): Promise<AddressResponse> | Observable<AddressResponse> | AddressResponse;

  createOrUpdateAddress(
    request: CreateAddressRequest,
  ): Promise<AddressResponse> | Observable<AddressResponse> | AddressResponse;
}

export function AddressServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["verifyAddress", "createOrUpdateAddress"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("AddressService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("AddressService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const ADDRESS_SERVICE_NAME = "AddressService";